requirements.txt
flask
line-bot-sdk
sqlalchemy
psycopg2-binary
apscheduler
requests
beautifulsoup4
gunicorn
